rl/requirements.txt
absl-py==0.8.1
alabaster==0.7.12
astor==0.8.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.2.0
Babel==2.7.0
backports.os==0.1.1
boto3==1.4.8
botocore==1.8.50
cached-property==1.5.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
colorama==0.4.1
coverage==4.5.4
cycler==0.10.0
decorator==4.4.0
dill==0.3.1.1
docutils==0.15.2
dss-python==0.10.1
filelock==3.0.12
funcsigs==1.0.2
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
gym==0.14.0
h5py==2.10.0
idna==2.8
imagesize==1.1.0
importlib-metadata==0.23
isort==4.3.15
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.1.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lxml==4.2.4
lz4==2.2.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.0
mccabe==0.6.1
more-itertools==7.2.0
Naked==0.1.31
networkx==2.2
nose2==0.8.0
numpy==1.16.2
numpydoc==0.9.1
opencv-python==4.1.1.26
OpenDSSDirect.py==0.3.7
opt-einsum==3.1.0
ortools==7.0.6546
packaging==19.2
pandas==0.25.1
path.py==12.0.1
pbr==5.1.3
plotly==1.9.6
pluggy==0.13.0
protobuf==3.8.0
psutil==5.6.3
py==1.8.0
pycparser==2.19
pyglet==1.3.2
Pygments==2.4.2
pylint==2.3.1
pymodbus==2.1.0
pyparsing==2.4.2
PyPrind==2.11.2
pyrsistent==0.15.4
pyserial==3.4
pytest==5.2.1
python-dateutil==2.7.3
pytz==2019.3
PyYAML==5.1
ray==0.7.5
redis==3.3.8
requests==2.21.0
s3transfer==0.1.13
scipy==1.1.0
setproctitle==1.1.10
six==1.11.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
testresources==2.0.1
typed-ast==1.3.1
urllib3==1.24.1
virtualenv==16.4.3
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.1
zipp==0.6.0
